chore: define store
